+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Release (2026-MM-DD)
 
+- `automation`:
+  - [v0.1.0](services/automation/CHANGELOG.md#v010)
+    - **New**: API for STACKIT Automation
+    - [Usage example](https://github.com/stackitcloud/stackit-sdk-go/tree/main/examples/automation)
 - `observability`:
   - [v0.25.0](services/observability/CHANGELOG.md#v0250)
     - `v1api`: